Move tests back to struct.test.ts
