+++ packages/typegpu/tests/struct.test.ts
@@ -18,7 +18,9 @@ import {
   vec3h,
   vec3u,
 } from '../src/data/index.ts';
+import tgpu from '../src/index.ts';
 import type { Infer } from '../src/shared/repr.ts';
+import { parse, parseResolved } from './utils/parseResolved.ts';
 
 describe('struct', () => {
   it('aligns struct properties when measuring', () => {
@@ -296,4 +298,84 @@ describe('struct', () => {
 
   //   expect(defaultStruct).toStrictEqual({ arr: [0] });
   // });
+
+  it('generates correct code when struct default constructor is used', () => {
+    const Nested = struct({ prop1: vec2f, prop2: u32 });
+    const Outer = struct({
+      nested: Nested,
+    });
+
+    const testFunction = tgpu.fn([])(() => {
+      const defaultValue = Outer();
+    });
+
+    expect(parseResolved({ testFunction })).toBe(parse(`
+          struct Nested {
+            prop1: vec2f,
+            prop2: u32,
+          }
+  
+          struct Outer {
+            nested: Nested,
+          }
+  
+          fn testFunction() {
+            var defaultValue = Outer();
+          }
+        `));
+  });
+
+  it('generates correct code when struct clone is used', () => {
+    const TestStruct = struct({
+      x: u32,
+      y: f32,
+    });
+
+    const testFn = tgpu.fn([])(() => {
+      const myStruct = TestStruct({ x: 1, y: 2 });
+      const myClone = TestStruct(myStruct);
+      return;
+    });
+
+    expect(parseResolved({ testFn })).toBe(
+      parse(`
+        struct TestStruct {
+          x: u32,
+          y: f32,
+        }
+  
+        fn testFn() {
+          var myStruct = TestStruct(1, 2);
+          var myClone = (myStruct);
+          return;
+        }`),
+    );
+  });
+
+  it('generates correct code when complex struct clone is used', () => {
+    const TestStruct = struct({
+      x: u32,
+      y: f32,
+    });
+
+    const testFn = tgpu.fn([])(() => {
+      const myStructs = [TestStruct({ x: 1, y: 2 })] as const;
+      const myClone = TestStruct(myStructs[0]);
+      return;
+    });
+
+    expect(parseResolved({ testFn })).toBe(
+      parse(`
+        struct TestStruct {
+          x: u32,
+          y: f32,
+        }
+  
+        fn testFn() {
+          var myStructs = array<TestStruct, 1>(TestStruct(1, 2));
+          var myClone = (myStructs[0]);
+          return;
+        }`),
+    );
+  });
 });
