+++ packages/typegpu/tests/tgsl/wgslGenerator.test.ts
@@ -620,86 +620,6 @@ describe('wgslGenerator', () => {
     expect((res.dataType as unknown as WgslArray).elementType).toBe(TestStruct);
   });
 
-  it('generates correct code when struct default constructor is used', () => {
-    const Nested = d.struct({ prop1: d.vec2f, prop2: d.u32 });
-    const Outer = d.struct({
-      nested: Nested,
-    });
-
-    const testFunction = tgpu.fn([])(() => {
-      const defaultValue = Outer();
-    });
-
-    expect(parseResolved({ testFunction })).toBe(parse(`
-        struct Nested {
-          prop1: vec2f,
-          prop2: u32,
-        }
-
-        struct Outer {
-          nested: Nested,
-        }
-
-        fn testFunction() {
-          var defaultValue = Outer();
-        }
-      `));
-  });
-
-  it('generates correct code when struct clone is used', () => {
-    const TestStruct = d.struct({
-      x: d.u32,
-      y: d.f32,
-    });
-
-    const testFn = tgpu.fn([])(() => {
-      const myStruct = TestStruct({ x: 1, y: 2 });
-      const myClone = TestStruct(myStruct);
-      return;
-    });
-
-    expect(parseResolved({ testFn })).toBe(
-      parse(`
-      struct TestStruct {
-        x: u32,
-        y: f32,
-      }
-
-      fn testFn() {
-        var myStruct = TestStruct(1, 2);
-        var myClone = (myStruct);
-        return;
-      }`),
-    );
-  });
-
-  it('generates correct code when complex struct clone is used', () => {
-    const TestStruct = d.struct({
-      x: d.u32,
-      y: d.f32,
-    });
-
-    const testFn = tgpu.fn([])(() => {
-      const myStructs = [TestStruct({ x: 1, y: 2 })] as const;
-      const myClone = TestStruct(myStructs[0]);
-      return;
-    });
-
-    expect(parseResolved({ testFn })).toBe(
-      parse(`
-      struct TestStruct {
-        x: u32,
-        y: f32,
-      }
-
-      fn testFn() {
-        var myStructs = array<TestStruct, 1>(TestStruct(1, 2));
-        var myClone = (myStructs[0]);
-        return;
-      }`),
-    );
-  });
-
   it('generates correct code for array expressions with derived elements', () => {
     const testFn = tgpu.fn([], d.f32)(() => {
       const arr = [derivedV2f.$, std.mul(derivedV2f.$, d.vec2f(2, 2))];
